Restore configuration properties resolution for OpenAI
